+++ CHANGELOG.md
@@ -56,6 +56,7 @@ Previous classification is not required if changes are simple or all belong to t
 - Changes in the prompt of `QuestionAnsweringPlugin` to enhance language detection in the response.
 - Slight improvement in the `DeleteAsync` method in `CosmosRepository`.
 - Some boy scouting and typo fixes in comments.
+- Added new extension method to add Qdrant and Azure Search AI as Keyed Memory Store.
 
 ## [8.1.4]